Add rememberme functionality
Two new options, `rememberMeField` and `rememberMeAdditionalMinutes`.

The first one sets which field to query for the rememberMe option from the
users request. This can be a nested field, like `auth.remember_me`.

The second field is the number of minutes that will be added to
the token expiration if rememberMe is `true`. With the default
token options this equals to 2 weeks.